chore(middleware): add validate-request
- add validateRequest middleware
- add validation helper
- modify whitelist routes
- modify test script
- add unit tests
- modify integration tests
