+++ src/middlewares/validate-request.ts
@@ -1,0 +1,12 @@
+import { Response, Request, NextFunction } from "express";
+import {validationResult} from 'express-validator';
+import { RequestValidationError } from "../errors/request-validation-error";
+
+export const validateRequest = (req: Request, res: Response, next: NextFunction) => {
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    throw new RequestValidationError(errors.array());
+  }
+  next();
+}
