+++ src/routes/whitelist.ts
@@ -1,33 +1,12 @@
 import express, { Request, Response } from "express";
-import { body, validationResult } from "express-validator";
 
 import { whitelists } from "../db/whitelist";
 import { WhitelistModel } from "../models/whitelist";
-import { RequestValidationError } from "../errors/request-validation-error";
 import { NotFoundError } from "../errors/not-found-error";
+import { validateRequest } from "../middlewares/validate-request";
+import { postValidatorObject } from "../helpers/validation";
 
 const router = express.Router();
-
-const postValidatorObject = [
-  body("path")
-    .trim()
-    .exists()
-    .isString()
-    .isLength({ min: 1 })
-    .withMessage("Path is required"),
-  body("ip")
-    .isString()
-    .matches(/^(?:[0-9]{1,3}\.){3}[0-9]{1,3}$/)
-    .withMessage("Must be a valid IP address"),
-];
-
-const validatePostValidatorBody = (req: express.Request) => {
-  const errors = validationResult(req);
-
-  if (!errors.isEmpty()) {
-    throw new RequestValidationError(errors.array());
-  }
-};
 
 router.get("/whitelists", (req: Request, res: Response) => {
   return res.status(200).send({
@@ -39,8 +18,8 @@ router.get("/whitelists", (req: Request, res: Response) => {
 router.post(
   "/whitelists",
   postValidatorObject,
+  validateRequest,
   (req: Request, res: Response) => {
-    validatePostValidatorBody(req);
     const { path, ip }: { path: string; ip: string } = req.body;
 
     const whitelist = whitelists[path];
@@ -61,13 +40,12 @@ router.post(
   }
 );
 
-router.post(
+router.delete(
   "/whitelists/remove",
   postValidatorObject,
+  validateRequest,
   (req: Request, res: Response) => {
-    validatePostValidatorBody(req);
     const { path, ip }: { path: string; ip: string } = req.body;
-
     const whitelist = whitelists[path];
 
     if (!whitelist) {
